-Se cargan cambios

+++ src/database/migrations/BaseMigration.php
@@ -1,9 +1,10 @@
 <?php
-namespace Crudvel\Migrations;
+namespace Crudvel\Database\Migrations;
 
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\DB;
 
 class BaseMigration extends Migration
 {
@@ -17,7 +18,11 @@ class BaseMigration extends Migration
 
     public function down()
     {
-        Schema::drop($this->mainTable);
+        if(Schema::hasTable($this->mainTable)){            
+            DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+            Schema::drop($this->mainTable);
+            DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+        }
     }
 
     public function makeStylesColumns($table){
